Make tag checks case-sensitive in a few formats that had it wrong
Tags could very well be case insensitive, but only if we see that in
a spec or in the wild and to the best of my knowledge this doesn't
apply to any of these formats.  Some of them even have their tags made
with our own *2john tools (which I then checked).

I also made sure nothing in our tree nor in the Test Suite has these
tags in a different case.

The only format that had me uncertain was sapH but some googling
seemed to confirm our tags for it are the correct (sometimes mixed)
case.

The LDAP tags I included here, {SSHA} and {SSHA512}, are not likely to
exist in lowercase anywhere while older ones such as {md5} or {sha}
theoretically could, but we only ever supported those in uppercase and
no-one ever complained.